Fix drag scrolling
When dragging items down, the list wouldn't scroll because the bottom
part of the list was hidden by the search box.  We now limit the size
of the box.

This also always maxes out the height of the popup.  This stopped an
issue I was seeing where occasionally the popup ended up being super
short (like maybe 100px).